Remove css and move file
