--- projects/packages/forms/src/blocks/contact-form/components/akismet-panel/styles.css
@@ -1,8 +1,0 @@
-.jetpack-akismet-panel__buttons {
-	display: flex;
-	gap: 8px;
-}
-
-.jetpack-akismet-panel__text {
-	margin-bottom: 16px;
-}
